v2.8.2

+++ SkEditor/Utilities/Projects/Elements/File.cs
@@ -1,4 +1,5 @@
 ﻿using CommunityToolkit.Mvvm.Input;
+using FluentAvalonia.UI.Controls;
 using SkEditor.API;
 using SkEditor.Utilities.Files;
 using System;
@@ -22,8 +23,7 @@ public class File : StorageElement
         Name = Path.GetFileName(file);
         IsFile = true;
 
-        var icon = Files.Icon.GetIcon(Path.GetExtension(file));
-        if (icon is not null) Icon = icon;
+        UpdateIcon();
 
         OpenInExplorerCommand = new RelayCommand(OpenInExplorer);
         DeleteCommand = new RelayCommand(DeleteFile);
@@ -36,8 +36,14 @@ public class File : StorageElement
         Process.Start(new ProcessStartInfo(Parent.StorageFolderPath) { UseShellExecute = true });
     }
 
-    public void DeleteFile()
+    public async void DeleteFile()
     {
+        var result = await SkEditorAPI.Windows.ShowDialog("Delete File", 
+            $"Are you sure you want to delete {Name} from the file system?",
+            icon: Symbol.Delete, primaryButtonText: "Delete", cancelButtonText: "Cancel", translate: false);
+
+        if (result != ContentDialogResult.Primary) return;
+
         System.IO.File.Delete(StorageFilePath);
         Parent.Children.Remove(this);
     }
@@ -61,10 +67,17 @@ public class File : StorageElement
         StorageFilePath = newPath;
         Name = newName;
 
+        UpdateIcon();
         RefreshSelf();
     }
 
     public override void HandleClick() => FileHandler.OpenFile(StorageFilePath);
+
+    public void UpdateIcon()
+    {
+        var icon = Files.Icon.GetIcon(Path.GetExtension(StorageFilePath));
+        if (icon is not null) Icon = icon;
+    }
 
     public void CopyAbsolutePath()
     {
