+++ SkEditor/Utilities/Projects/Elements/Folder.cs
@@ -1,6 +1,10 @@
-﻿using Avalonia.Threading;
+﻿using Avalonia.Controls;
+using Avalonia.Threading;
 using CommunityToolkit.Mvvm.Input;
+using FluentAvalonia.UI.Controls;
+using Serilog;
 using SkEditor.API;
+using SkEditor.Controls.Sidebar;
 using SkEditor.Utilities.Files;
 using SkEditor.Views;
 using SkEditor.Views.Projects;
@@ -9,6 +13,7 @@ using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using static SkEditor.Controls.Sidebar.ExplorerSidebarPanel;
 
 namespace SkEditor.Utilities.Projects.Elements;
 
@@ -18,12 +23,13 @@ public class Folder : StorageElement
 
     public Folder(string folder, Folder? parent = null)
     {
-        folder = Uri.UnescapeDataString(folder).FixLinuxPath();
+        folder = Uri.UnescapeDataString(folder).FixLinuxPath().TrimEnd(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar);
 
         Parent = parent;
         StorageFolderPath = folder;
         Name = Path.GetFileName(folder);
         IsFile = false;
+        IsRootFolder = parent is null;
 
         Children = [];
         LoadChildren();
@@ -34,6 +40,7 @@ public class Folder : StorageElement
         CopyAbsolutePathCommand = new RelayCommand(CopyAbsolutePath);
         CreateNewFileCommand = new RelayCommand(() => CreateNewElement(true));
         CreateNewFolderCommand = new RelayCommand(() => CreateNewElement(false));
+        CloseProjectCommand = new RelayCommand(CloseProject);
     }
 
     private async void LoadChildren()
@@ -50,10 +57,26 @@ public class Folder : StorageElement
         Process.Start(new ProcessStartInfo(StorageFolderPath) { UseShellExecute = true });
     }
 
-    public void DeleteFolder()
+    public async void DeleteFolder()
     {
+        var result = await SkEditorAPI.Windows.ShowDialog("Delete File",
+            $"Are you sure you want to delete {Name} from the file system?",
+            icon: Symbol.Delete, primaryButtonText: "Delete", cancelButtonText: "Cancel", translate: false);
+
+        if (result != ContentDialogResult.Primary) return;
+
         Directory.Delete(StorageFolderPath, true);
-        Parent.Children.Remove(this);
+        Parent?.Children?.Remove(this);
+        if (Parent is null) CloseProject();
+    }
+
+    private static void CloseProject()
+    {
+        ProjectOpener.FileTreeView.ItemsSource = null;
+        Folder? ProjectRootFolder = null;
+        ExplorerPanel? Panel = Registries.SidebarPanels.FirstOrDefault(x => x is ExplorerPanel) as ExplorerPanel;
+        StackPanel NoFolderMessage = Panel.Panel.NoFolderMessage;
+        NoFolderMessage.IsVisible = ProjectRootFolder == null;
     }
 
     public override void RenameElement(string newName, bool move = true)
